refact: navtop, outlet components, footer layout refactoring
