+++ src/components/Footer.js
@@ -1,0 +1,12 @@
+import styled from "styled-components";
+
+const Footer = () => {
+    return (
+        <div>
+            <h1>footer</h1>
+            <p>footer blahblah</p>
+        </div>
+    )
+};
+
+export default Footer
